Refactor out PinUvAuthToken-stuff to make webauthn.rs a bit easier to handle
